requirements.txt
fastapi
uvicorn
pydantic
numpy
openai
openenv-core
requests
pandas
